+++ NEWS.md
@@ -1,8 +1,12 @@
 # tidycovid19 0.0.0.9000
 
+* `*_covid19_spread()`: Added option to exclude unhighlighted countries
+
 * `shiny_covid19_spread()`: Changed API to include all plot options
 
 * `*_covid19_spread()`: Added option for averaged daily changes
+
+* Included Oxford NPI data
 
 * `*_covid19_spread()`: Added option for linear Y-axis